Added softmax and crossentropy loss so that Paquito is able to classify digits from the MNIST dataset

+++ include/types.h
@@ -4,7 +4,7 @@
 #include <cmath>
 #include <vector>
 
-typedef double myType;
+typedef float myType;
 typedef std::vector<std::vector<myType>> Matrix;
 typedef std::vector<myType> Vector;
